OIC, CBOR support

+++ ManuvrOS/Types/cbor-cpp/decoder.cpp
@@ -52,7 +52,7 @@ void decoder::run() {
                 switch(majorType) {
                     case 0: // positive integer
                         if(minorType < 24) {
-                            _listener->on_integer(minorType);
+                            _listener->on_integer((unsigned int)minorType);
                         } else if(minorType == 24) { // 1 byte
                             _currentLength = 1;
                             _state = STATE_PINT;
@@ -72,7 +72,7 @@ void decoder::run() {
                         break;
                     case 1: // negative integer
                         if(minorType < 24) {
-                            _listener->on_integer(-1 -minorType);
+                            _listener->on_integer((int)-1 -minorType);
                         } else if(minorType == 24) { // 1 byte
                             _currentLength = 1;
                             _state = STATE_NINT;
